apps/worker/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
playwright
python-telegram-bot==20.7
python-dotenv==1.0.0
httpx
aiosqlite
apscheduler
google-generativeai

# Anti-detection stealth
playwright-stealth==1.0.6

# Content Sources
beautifulsoup4
python-dateutil
arxiv
pytrends
aiohttp
tweepy
